refactory structure files
