+++ src/views/pages/Home/Home.ejs
@@ -1,8 +1,8 @@
 <%- include('./components/HomeHero') %>
 <div class="container mx-auto bg-gray-900 pt-12 pb-12">
     <div class="flex flex-col space-y-20">
-        <%- include('../components/ListHomeSection', { data: data.movies, type: "movie" }) %>
-        <%- include('../components/ListHomeSection', { data: data.series, type: "serie" }) %>
-        <%- include('../components/ListHomeSection', { data: data.genres, type: "genre" }) %>
+        <%- include('../../components/ListHomeSection', { data: data.movies, type: "movie" }) %>
+        <%- include('../../components/ListHomeSection', { data: data.series, type: "serie" }) %>
+        <%- include('../../components/ListHomeSection', { data: data.genres, type: "genre" }) %>
     </div>
 </div>
